style: type casted.

+++ src/ngx_http_small_light_imagemagick.c
@@ -46,8 +46,8 @@ void ngx_http_small_light_imagemagick_adjust_image_offset(ngx_http_request_t *r,
         return;
     }
 
-    sz->sx = x;
-    sz->sy = y;
+    sz->sx = (double) x;
+    sz->sy = (double) y;
 }
 
 ngx_int_t ngx_http_small_light_imagemagick_init(ngx_http_request_t *r, ngx_http_small_light_ctx_t *ctx)
